Imported face-tracking demo from OpenCVproj.py into CameraInput.py and made coordinate yielder
used the Demo logic to create a Generator function "trackObject()" which YIELDs an (x,y) tuple of the Bounding Box' location.

main.py also calls this new function and prints to terminal to give live-time coordinates of the Object of Interest.

In later version, should switch out the faceCascade for a Shapes cascade