ColorLabeler.label() now also returns RGB value of contour
So apparently the code example david imported was using L*a*b* the whole time which made extracting objects' color values both confusing and difficult as hell for me for the past 2 days. What made this all worse is that opencv formats their LAB as (0-255, 0-256, 0-256) instead of the standard (0-100, -128~128, -128~128) so I needed to import some colormath module functionality to convert L*a*b* --> RGB. I lost a lot of sleep for this.